+++ src/commands/includes/removeJobsByMaxAge.lua
@@ -3,6 +3,7 @@
 ]]
 
 -- Includes
+--- @include "batches"
 --- @include "removeJob"
 
 local function removeJobsByMaxAge(timestamp, maxAge, targetSet, prefix, maxLimit)
@@ -11,5 +12,13 @@ local function removeJobsByMaxAge(timestamp, maxAge, targetSet, prefix, maxLimit
   for i, jobId in ipairs(jobIds) do
     removeJob(jobId, false, prefix, false --[[remove debounce key]])
   end
-  rcall("ZREMRANGEBYSCORE", targetSet, "-inf", start)
+  if #jobIds > 0 then
+    if #jobIds < maxLimit then
+      rcall("ZREMRANGEBYSCORE", targetSet, "-inf", start)
+    else
+      for from, to in batches(#jobIds, 7000) do
+        rcall("ZREM", targetSet, unpack(jobIds, from, to))
+      end
+    end
+  end
 end